+++ cargo-e/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.11](https://github.com/davehorner/cargo-e/compare/cargo-e-v0.1.10...cargo-e-v0.1.11) - 2025-03-15
+
+### Added
+
+- implement self-update functionality for Windows (spawn batch) in
+
 ## [0.1.10](https://github.com/davehorner/cargo-e/compare/cargo-e-v0.1.9...cargo-e-v0.1.10) - 2025-03-15
 
 ### Added